wsgi: Pass worker pids to new manager when reloading
This allows the new manager to recognize them as they exit instead of
logging

   Ignoring wait() result from unknown PID ...

With the improved knowledge of subprocesses, we can kick the log level
for the above message up to warning; we no longer expect it to trigger
in practice.

Additionally, add a new tunable, `stale_worker_timeout`. Once this many
seconds pass following a reload, the manager process will issue SIGKILLs
to the remaining stale workers. This ensures there is some limit to how
long old code and configs may still be running.

Drive-by: Add logging to ServersPerPortStrategy.register_worker_exit
that's comparable to what WorkersStrategy does.

Change-Id: I8227939d04fda8db66fb2f131f2c71ce8741c7d9