WIP: Reconstructor: Use past node and abort to handoff
Before the reconstructor makes a bunch of connections to go rebuild
the missing fragment it makes it takes advantage of the last_primary
table in the ring. If there is a last primariy for the missing frags
partition then it'll first do a HEAD to see if the missing fragment
for the same timestamp exists there. If it does, it will tell the
current ssync session to abort and skip this datafile, and let the
handoff REVERT the missing fragment saving a needless rebuild.

In order to abort the current diskfile in ssync, a new ssync exception
has been created called SsyndAbortDataFile which can be thrown from the
sync_diskfile_builder callback. When thrown the ssync sender will move
onto the next diskfile to send.

Finally it's also triggering a statsd found.last_node metric because it
might be interesting to monitor.

Change-Id: Ie6e039a52067db1c02febb39f9c5c0829e5ed0a4