ring: Keep track of last primary nodes from last rebalance
This patch adds an additional part2dev array to the ring, which is
plumbed through the ring, ring_data and builder objects. Note that this
may only be serialized to ring.gz files if using ring-format version 2.

Because we only ever move 1 partition replica at a time during a
rebalance we can store the last node that was the primary for a
partition before the last rebalance.
This will allow us to be smarter in places like the reconstructor and
replicators as we'll know where to look for a missing replica/fragment
after a rebalance.

The array is filled with NONE_DEVs at the start of a rebalance, and then
an old device is placed inside when wanting assigning new partitions.

A ring method `get_last_part_node(part)` now exists which will return
the last primary node for a part or None if there wasn't any. The latter
can occur if a device is removed or a part didn't move during the last
rebalance.

This patch also adds a 'for_read' optional parameter for get_more_nodes
which when true will add the last primary to the front of the handoff
list, if there is is one.

Change-Id: Iaa66b9fe57c924445a5ede908053e78fc0ed8817