container: Add delimiter-depth query param
This allows container listings to have a semantic similar to running
`find <dir> -maxdepth <N>`, making it easier to get a high-level view of
the contents of a deeply-nested container without resorting to
potentially thousands of requests to perform a depth-first traversal.

Limit depth values to the range [1, MAX_DELIMITER_DEPTH] to prevent
abuse; values outside of that range will be silently changed to the
default value (1). MAX_DELIMITER_DEPTH defaults to 5 but may be
configured via the [swift-constraints] section of swift.conf.

Change-Id: I43f68d7bef84296efb3abf1c7b3c6533d877393b