ring: Add a rebalance history in the ring
This patch introduces 2 new datastructures to the ring:
 - _past2part2dev_id
 - _past2part2index

Both are like the replica2part2dev_id structure except it stores instead
of replicas each index that many rebalances ago.

  ring._past2part2dev_id[0][50]

Would grab the device id of the primary that moved for part 50 in the
last rebalance.

  ring._past2part2dev_id[2][50]

Would grab the same but 3 rebalances ago. If nothing moved for that part
we'd get a NONE_DEV.

The _past2part2index tracks the index that the corrisponding node was at
when it was a primary. And old version of this patch stored them
together but that only allowed us to store the last primary. Storing
this way gives us a history.
By default the history is 1, so in reality acts like a single last part
table. But there is a `_history_count` parameter in the ring which
controls the maximum amount of history (in rebalances) to store.

These have been plumbed through the ring, ring_data and builder objects.
And it can only be serialized to ring.gz files if using ring-format version 2.

Having access to the old indexes allows us to be smarter in places like
the reconstructor as we'll know where to look for a missing replica/fragment
after a rebalance.

Change-Id: I1e605a10cf4b586a1e51d8fda680bd8bc6e5ff85