Remove hard dependency on netifaces
The project was archived in 2021, and we can fairly easily replace it
with some ctypes code to call getifaddrs ourselves.

Be willing to fall back to netifaces (with a warning) in case getifaddrs
is not available, but I'm fairly certain it will be for all platforms we
support.

While we're at it, pull a bunch of ip-address-related functions into a
new module under utils.

Partial-Bug: #2019233
Change-Id: I1189a60204cf96c291619f8d8ec957ed8a5be1ce